removed test output and now passes tests
