+++ internal/debugger/spatialite/go-spatialite/spatialite.go
@@ -4,7 +4,7 @@ import (
 	"database/sql"
 	"errors"
 
-	"github.com/mattn/go-sqlite3"
+	sqlite3 "github.com/mattn/go-sqlite3"
 )
 
 type entrypoint struct {
