--- planar/triangulate/debugger/spatialite/go-spatialite/README.md
@@ -1,8 +1,0 @@
-# Go Spatialite
-[![godoc](http://img.shields.io/badge/godoc-reference-blue.svg?style=flat)](https://godoc.org/gopkg.in/shaxbee/go-spatialite.v0) 
-[![license](http://img.shields.io/badge/license-MIT-blue.svg?style=flat)](https://raw.githubusercontent.com/shaxbee/go-snowflake/master/LICENSE) 
-
-Spatialite SQL Driver for Golang.
-
-copied from : github.com/shaxbee/go-spatialite
-
