+++ planar/triangulate/subdivision/subdivision.go
@@ -7,7 +7,7 @@ import (
 	"log"
 	"sort"
 
-	"github.com/go-spatial/geom/planar/triangulate/debugger"
+	"github.com/go-spatial/geom/internal/debugger"
 	"github.com/go-spatial/geom/planar/triangulate/geometry"
 	"github.com/go-spatial/geom/planar/triangulate/quadedge"
 	"github.com/go-spatial/geom"
